bpf, security: Add Checmate
This adds the minor LSM Checmate. The purpose of Checmate is to act as an
extensible LSM in which you can load security modules. The module has a
simple API, as it's meant to have most of the logic in BPF hooks. It has
three APIs that are accessible via prctl.

As follows:
* Install hook: This appends a new BPF program to a given hook. Hook
		programs themselves must be unique BPF programs.
* Reset hook: This detaches all bpf programs asssociated with a hook.
* Deny Reset: This locks a hook, preventing reset. In production
	      operation, it's expected that the user would lock
	      their hooks.

At least while this is in early days, I think it makes sense to
restrictChecmate to loading programs that have been compiled with
the current kernel ABI. We can further stabilize the ABI, and
perhaps lift this restriction later.

Signed-off-by: Sargun Dhillon <sargun@sargun.me>