net/core: Allocate cgroup sk data before calling security_sk_(alloc/free)
This patch changes the order of allocations / calls to allocate the skcd
before calling into the LSM security subsystem. This is because if a
particular LSM uses a generic mechanism for getting the cgroup for a
given struct sock (fetching the skcd), it makes it difficult.

Also, since is applies to security_sk_free, on the LSM side it's
difficult to determine whether to look at skcd, or current.